+++ metricprofiler/src/main/java/com/expedia/adaptivealerting/metricprofiler/source/DefaultProfileSource.java
@@ -15,20 +15,58 @@
  */
 package com.expedia.adaptivealerting.metricprofiler.source;
 
+import com.expedia.adaptivealerting.anomdetect.util.AssertUtil;
+import com.expedia.adaptivealerting.anomdetect.util.HttpClientWrapper;
 import com.expedia.metrics.MetricDefinition;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import lombok.val;
+import org.apache.http.client.fluent.Content;
 
-import javax.validation.constraints.NotNull;
+import java.io.IOException;
+
+import static com.expedia.adaptivealerting.anomdetect.util.AssertUtil.isTrue;
 
 @RequiredArgsConstructor
 public class DefaultProfileSource implements ProfileSource {
 
-    @NotNull
-    private final MetricProfilerClient client;
+    @NonNull
+    private final HttpClientWrapper httpClient;
+
+    @NonNull
+    private final ObjectMapper objectMapper;
+
+    @NonNull
+    private final String baseUri;
+
+    public static final String FIND_DOCUMENT_PATH = "/api/metricProfiling/search/findByTags";
 
     @Override
     public Boolean profileExists(MetricDefinition metricDefinition) {
-        return client.profileExists(metricDefinition);
-    }
+        AssertUtil.notNull(metricDefinition, "metricDefinition can't be null");
+        val tags = metricDefinition.getTags().getKv();
 
+        isTrue(tags.size() > 0, "tags must not be empty");
+
+        val uri = baseUri + FIND_DOCUMENT_PATH;
+        Content content;
+        try {
+            val body = objectMapper.writeValueAsString(tags);
+            content = httpClient.post(uri, body);
+        } catch (IOException e) {
+            val message = "IOException while finding matching metrics for" +
+                    ": tags=" + tags +
+                    ", httpMethod=POST" +
+                    ", uri=" + uri;
+            throw new RuntimeException(message, e);
+        }
+        try {
+            objectMapper.readValue(content.asBytes(), Boolean.class);
+            return true;
+        } catch (IOException e) {
+            val message = "IOException while finding finding matching profile: tags=" + tags;
+            throw new RuntimeException(message, e);
+        }
+    }
 }
